v1.1.7
- Added Events
- Added isOnline()
- Added database ping
more fixes and improvements

+++ test/index.js
@@ -1,46 +1,59 @@
 const db = require("../lib/index"),
   url = process.argv[2].replace("--", "");
 
+
 async function test() {
   if (!url) {
     throw new Error(
       "A URL is required, for example node test/index.js --mongodb+srv://user:pass@cluster.domain.mongodb.net/database. Need Help ? Visit pogy.xyz/support"
     );
   }
-  const database = await db.connect(url);
+
+  const database = await db.connect(url, {
+    cache: true,
+  });
   if (database) {
     console.log("Starting tests");
     let schema = await new db.table("npm-test");
-    console.log("Created schema npm-test");
 
-    await schema.set("test.test", {
-      test: "test",
-    });
-    console.log("Added test to schema with value object {test: 'test'}");
+    if (schema) {
+      console.log("Created schema npm-test");
 
-    await schema.push("test.tests", "test");
-    console.log("Pushed test to tests array");
+      await schema.set("test.test", {
+        test: "test",
+      });
+      console.log("Added test to schema with value object {test: 'test'}");
 
-    await schema.add("test.testCount", 1);
-    console.log("Added 1 to testCount");
+      await schema.push("test.tests", "test");
+      console.log("Pushed test to tests array");
 
-    await schema.subtract("test.testCount", 1);
-    console.log("Subtracted 1 from testCount");
+      await schema.add("test.testCount", 1);
+      console.log("Added 1 to testCount");
 
-    await schema.get("test.tests");
-    console.log("Got tests array");
+      await schema.subtract("test.testCount", 1);
+      console.log("Subtracted 1 from testCount");
 
-    await schema.has("test.tests");
-    console.log("Checked if tests array exists");
+      await schema.get("test.tests");
+      console.log("Got tests array");
 
-    await schema.delete("test.tests");
-    console.log("Deleted tests array");
+      await schema.has("test.tests");
+      console.log("Checked if tests array exists");
 
-    await schema.all();
-    console.log("Got all data");
+      await schema.all();
+      console.log("Got all data");
 
-    await schema.drop();
-    console.log("Dropped schema");
+      const ping = await db.ping({
+        tableName: "npm-test",
+        dataToGet: "test.tests",
+      });
+
+      console.log("Got ping!", ping);
+
+      await schema.drop();
+      console.log("Dropped schema");
+    } else {
+      console.log("Database is offline");
+    }
   }
 }
